Rename networks to providers
